refactor: 마크다운 시리얼라이저 SOLID 원칙에 따라 파일 분리
- 896줄 단일 enum을 9개 파일로 책임 분리 (SRP)
- MarkdownSerializer는 공개 API만 위임하는 facade로 축소
- 타입 정의(Types), 정규식(Regex), 이스케이프(Escaping), 속성 빌더(AttributeBuilder) 분리
- 블록/인라인 직렬화·파싱을 각각 전용 모듈로 분리
- HTML 감지 로직을 독립 모듈로 추출
- 기존 공개 API 시그니처 100% 유지로 호출부 영향 없음
